# docs/swagger.yaml
definitions:
  api_helper.CreateIPRequest:
    properties:
      expirationAt:
        type: string
      ip-address:
        type: string
      type:
        type: integer
    required:
    - ip-address
    type: object
  api_helper.CreateSqlInjectRequest:
    properties:
      rule:
        type: string
    required:
    - rule
    type: object
  api_helper.CreateXssDetectRequest:
    properties:
      rule:
        type: string
    required:
    - rule
    type: object
  api_helper.LoginRequest:
    properties:
      captcha:
        type: string
      captchaId:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - captcha
    - captchaId
    - password
    - username
    type: object
  api_helper.Response:
    properties:
      data:
        type: object
      msg:
        type: string
      status:
        type: integer
    type: object
  api_helper.UpdateIPRequest:
    properties:
      expiration_at:
        type: string
      ip_address:
        type: string
    required:
    - ip_address
    type: object
  config.Config:
    properties:
      rateLimiter:
        description: 限速器配置
        properties:
          banCounter:
            description: 计数器常规值，超过触发常规封禁时间
            type: integer
          banDuration:
            description: 常规封禁时间，单位秒
            type: integer
          fixedWindow:
            description: 固定窗口模式配置
            properties:
              maxRequest:
                description: 窗口内最大请求数
                type: integer
              windowSize:
                description: 窗口大小,单位秒
                type: integer
            type: object
          leakyBucket:
            description: 漏桶模式配置
            properties:
              capacity:
                description: 桶容量
                type: integer
              leakyPerSecond:
                description: 漏水速度，单位秒
                type: integer
            type: object
          maxCounter:
            description: 计数器最大值，超过该值拉入永久IP黑名单
            type: integer
          mode:
            description: 限速算法，1:令牌桶2:漏桶3:固定窗口4:滑动窗口
            type: integer
          slideWindow:
            description: 滑动窗口模式配置
            properties:
              maxRequest:
                description: 窗口内最大队列数
                type: integer
              windowSize:
                description: 窗口大小，单位秒
                type: integer
            type: object
          tokenBucket:
            description: 令牌桶模式配置
            properties:
              maxToken:
                description: 最大令牌数
                type: integer
              tokenPerSecond:
                description: 令牌生成速度，单位秒
                type: integer
            type: object
        type: object
      secret:
        properties:
          jwtSecretKey:
            type: string
          sessionSecretKey:
            type: string
        type: object
      server:
        properties:
          targetAddress:
            type: string
        type: object
    type: object
  types.AddRoutingRequest:
    properties:
      method:
        type: string
      route:
        type: string
      type:
        type: integer
    required:
    - method
    - route
    - type
    type: object
info:
  contact: {}
  description: Golang编写的一款基于反向代理模式的web防火墙应用 By supercat0867
  license: {}
  title: GoWAFer
  version: v0.1
paths:
  /waf/api/v1/auth/dologin:
    post:
      consumes:
      - application/json
      description: dologin
      parameters:
      - description: 请求体
        in: body
        name: api_helper.LoginRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 处理登录
      tags:
      - Auth
  /waf/api/v1/auth/getCaptcha:
    get:
      consumes:
      - application/json
      description: 获取图片验证码
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 获取图片验证码
      tags:
      - Auth
  /waf/api/v1/config:
    get:
      description: 获取配置信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 获取配置信息
      tags:
      - Config
    put:
      description: 修改配置文件并重启服务
      parameters:
      - description: 请求体
        in: body
        name: config.Config
        required: true
        schema:
          $ref: '#/definitions/config.Config'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 修改配置文件
      tags:
      - Config
  /waf/api/v1/ip:
    get:
      description: 分页查询IP
      parameters:
      - description: 关键词
        in: query
        name: keywords
        type: string
      - description: IP类型
        in: query
        name: type
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页面大小
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 分页查询IP
      tags:
      - IP
    post:
      consumes:
      - application/json
      description: 新增IP记录
      parameters:
      - description: 请求体
        in: body
        name: api_helper.CreateIPRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.CreateIPRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 新增IP记录
      tags:
      - IP
  /waf/api/v1/ip/{id}:
    delete:
      description: 删除IP
      parameters:
      - description: IP主键ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 删除IP
      tags:
      - IP
    patch:
      description: 编辑IP
      parameters:
      - description: IP主键ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求体
        in: body
        name: api_helper.UpdateIPRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.UpdateIPRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 编辑IP
      tags:
      - IP
  /waf/api/v1/log:
    get:
      description: 查询指定天数和小时数的日志记录
      parameters:
      - description: 查询范围天数
        in: query
        name: days
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 查询指定天数和小时数的日志记录
      tags:
      - Log（日志模块）
  /waf/api/v1/log/getBlockLog:
    get:
      description: 分页查询被拦截的流量日志
      parameters:
      - description: 查询IP
        in: query
        name: keywords
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页面大小
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 分页查询被拦截的流量日志
      tags:
      - Log（日志模块）
  /waf/api/v1/routing:
    delete:
      consumes:
      - application/json
      description: 删除路由管理记录
      parameters:
      - description: 请求体
        in: body
        name: types.AddRoutingRequest
        required: true
        schema:
          $ref: '#/definitions/types.AddRoutingRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 删除路由管理记录
      tags:
      - Routing
    get:
      description: 分页查询路由管理记录
      parameters:
      - description: 关键词
        in: query
        name: keywords
        type: string
      - description: 路由类型
        in: query
        name: type
        required: true
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页面大小
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 分页查询路由管理记录
      tags:
      - Routing
    post:
      consumes:
      - application/json
      description: 新增路由管理记录
      parameters:
      - description: 请求体
        in: body
        name: types.AddRoutingRequest
        required: true
        schema:
          $ref: '#/definitions/types.AddRoutingRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 新增路由管理记录
      tags:
      - Routing
  /waf/api/v1/sqlInject:
    get:
      description: 分页查询SQL注入规则
      parameters:
      - description: 关键词
        in: query
        name: keywords
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页面大小
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 分页查询SQL注入规则
      tags:
      - SQlInject
    post:
      consumes:
      - application/json
      description: 新增sql注入规则
      parameters:
      - description: 请求体
        in: body
        name: api_helper.CreateSqlInjectRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.CreateSqlInjectRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 新增sql注入规则
      tags:
      - SQlInject
  /waf/api/v1/sqlInject/{id}:
    delete:
      description: 删除sql注入规则
      parameters:
      - description: 主键ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 删除sql注入规则
      tags:
      - SQlInject
    patch:
      description: 编辑sql注入规则
      parameters:
      - description: 主键ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求体
        in: body
        name: api_helper.CreateSqlInjectRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.CreateSqlInjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 编辑sql注入规则
      tags:
      - SQlInject
  /waf/api/v1/xssDetect:
    get:
      description: 分页查询xss攻击匹配规则
      parameters:
      - description: 关键词
        in: query
        name: keywords
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 页面大小
        in: query
        name: perPage
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 分页查询xss攻击匹配规则
      tags:
      - XssDetect
    post:
      consumes:
      - application/json
      description: 新增xss攻击匹配规则
      parameters:
      - description: 请求体
        in: body
        name: api_helper.CreateXssDetectRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.CreateXssDetectRequest'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 新增xss攻击匹配规则
      tags:
      - XssDetect
  /waf/api/v1/xssDetect/{id}:
    delete:
      description: 删除sql注入规则
      parameters:
      - description: 主键ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 删除sql注入规则
      tags:
      - XssDetect
    patch:
      description: 编辑xss攻击匹配规则
      parameters:
      - description: 主键ID
        in: path
        name: id
        required: true
        type: integer
      - description: 请求体
        in: body
        name: api_helper.CreateXssDetectRequest
        required: true
        schema:
          $ref: '#/definitions/api_helper.CreateXssDetectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_helper.Response'
      summary: 编辑xss攻击匹配规则
      tags:
      - XssDetect
swagger: "2.0"